Update to the new babel ASTs and conver to TypeScript
This converts the package into a type safe AST walker with only slightly reduced functionality, while also updating it to use the newer babel ASTs.

It drops support for node < 10 (as these are no longer officially supported). A nice side effect of this is that we no longer require `@babel/runtime`